context-keeper/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
fastapi
uvicorn
